[Refactor] #36 기타 오류 수정
