+++ src/main/java/com/umc9th/areumdap/domain/device/controller/DeviceController.java
@@ -1,8 +1,9 @@
-package com.umc9th.areumdap.domain.device.controller.docs;
+package com.umc9th.areumdap.domain.device.controller;
 
 import com.umc9th.areumdap.common.response.ApiResponse;
 import com.umc9th.areumdap.common.status.SuccessStatus;
-import com.umc9th.areumdap.domain.device.dto.reqeust.RegisterDeviceRequest;
+import com.umc9th.areumdap.domain.device.controller.docs.DeviceControllerDocs;
+import com.umc9th.areumdap.domain.device.dto.request.RegisterDeviceRequest;
 import com.umc9th.areumdap.domain.device.service.DeviceCommandService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -10,16 +11,15 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-@CrossOrigin(origins = "http://127.0.0.1:5500") // TODO: 환경변수 적용
 @RequestMapping("/api/device")
 @RestController
 @RequiredArgsConstructor
-public class DeviceController implements DeviceControllerDocs{
+public class DeviceController implements DeviceControllerDocs {
 
     private final DeviceCommandService deviceCommandService;
 
     @Override
-    @PostMapping("/tokens")
+    @PostMapping("/")
     public ResponseEntity<ApiResponse<Void>> registerDevice(
             @AuthenticationPrincipal Long userId,
             @Valid  @RequestBody RegisterDeviceRequest registerDeviceRequest
