+++ src/main/java/com/umc9th/areumdap/domain/device/service/DeviceQueryService.java
@@ -1,4 +1,11 @@
 package com.umc9th.areumdap.domain.device.service;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
 public class DeviceQueryService {
 }
